remove EINTR handling and add test if we are in main thread before setting SIGCHLD
